List All Authors in the Dataset Catalog

Retrieves a list of all authors who have contributed datasets to the catalog by fetching datasets and extracting unique author information.

Usage

get_authors(max_datasets = 1000)

Arguments

max_datasets

Numeric. Maximum number of datasets to retrieve for author extraction, defaults to 1000.

Value

A tibble (data frame) with the following columns:

name

Character. Name of the author.

count

Numeric. Number of datasets contributed by this author.

Examples

try({
  authors <- get_authors(max_datasets = 100)
  head(authors)
})
#> INFO [2025-03-18 13:07:30] Fetching datasets to extract author information...
#> INFO [2025-03-18 13:07:31] Successfully extracted 9 unique authors from 100 datasets
#>                                                                     name count
#> 1 Ministère de l'Agriculture, des Ressources Hydrauliques et de la Pêche    81
#> 2                                                     Habib Ben Boubaker     6
#> 3                                                                 rhaimi     4
#> 4                                                            Nawel Ajabi     2
#> 5                                                                 Rhaimi     2
#> 6                                                    abdessatar talhaoui     2