List Organizations

Retrieves organizations data from the Tunisian data catalog API (data.gov.tn) using faceted search. This function returns organizations that have published datasets.

Usage

get_organizations(min_count = 1)

Arguments

min_count

Integer. Minimum number of datasets an organization must have to be included in results. Default is 1, meaning only organizations with at least one dataset are returned.

Value

A tibble (data frame) with the following columns:

name

Character. Machine-readable name/identifier of the organization.

display_name

Character. Human-readable name of the organization.

dataset_count

Integer. Number of datasets published by the organization.

Examples

# \donttest{
try({
  # Get all organizations with at least 5 datasets
  orgs <- get_organizations(min_count = 5)
  head(orgs)
})
#> INFO [2025-03-07 00:49:46] Retrieving organizations using faceted search
#> SUCCESS [2025-03-07 00:49:46] Found 88 organizations via faceted search
#> # A tibble: 6 × 3
#>   name                               display_name                  dataset_count
#>   <chr>                              <chr>                                 <int>
#> 1 ministere-des-affaires-culturelles Minstère des affaires cultur…           284
#> 2 ministere                          Ministère de l'Industrie, de…           222
#> 3 1kelibia                           Commune de Kélibia                       95
#> 4 jammel                             Commune de Jemmel                        90
#> 5 enfidha                            Commune d'Enfidha                        87
#> 6 agareb                             Municipalité Agareb                      72
# }